+++ isy-security/CHANGELOG.md
@@ -1,5 +1,7 @@
 # 3.0.0
 
+- `IFS-2561`: Die Konfiguration via `rollenrechte.xml` erfolgt optional. 
+    - Initialisierung mit Standardwerten
 - `IFS-1852`: Bereitstellung einer Implementierung zur Authentifizierung eines Clients mit Client Credentials
     - Property `isy.security.rolesClaimName` eingeführt
 - `IFS-2302`: Umsetzung des Mappings von Rollen auf Rechte
@@ -14,3 +16,4 @@
     - Vereinheitlichung der Methodensignaturen
     - Ergänzung des BHKNZ als optionalen Parameter in allen Methoden (wird aktuell nur für den Resource Owner Password Credentials Flow ausgewertet)
     - Überarbeitung der Dokumentation und Parameterbenennung um Unklarheiten zu beseitigen
+- `ISY-83`: Annotation zur Authentifizierung innerhalb von Methoden hinzugefügt